LRQA-42394 Create a base portal release job to be extended by patch jobs
