+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/PortalReleaseJob.java
@@ -22,62 +22,28 @@ import java.util.Set;
 /**
  * @author Michael Hashimoto
  */
-public class PortalReleaseJob
-	extends BaseJob implements BatchDependentJob, PortalTestClassJob {
+public class PortalReleaseJob extends BasePortalReleaseJob {
 
 	public PortalReleaseJob(String jobName, String portalBranchName) {
-		super(jobName);
+		super(jobName, portalBranchName);
 
-		_portalBranchName = portalBranchName;
-
-		_jenkinsGitWorkingDirectory =
-			JenkinsResultsParserUtil.getJenkinsGitWorkingDirectory();
-
-		_portalGitWorkingDirectory =
-			JenkinsResultsParserUtil.getPortalGitWorkingDirectory(
-				portalBranchName);
+		GitWorkingDirectory jenkinsGitWorkingDirectory =
+			getJenkinsGitWorkingDirectory();
 
 		jobProperties.putAll(
 			JenkinsResultsParserUtil.getProperties(
 				new File(
-					_portalGitWorkingDirectory.getWorkingDirectory(),
-					"test.properties")));
-
-		jobProperties.putAll(
-			JenkinsResultsParserUtil.getProperties(
-				new File(
-					_jenkinsGitWorkingDirectory.getWorkingDirectory(),
+					jenkinsGitWorkingDirectory.getWorkingDirectory(),
 					"commands/dependencies/test-portal-release.properties")));
 	}
 
 	@Override
 	public Set<String> getBatchNames() {
-		String testBatchNamesString = JenkinsResultsParserUtil.getProperty(
-			jobProperties, "test.batch.names[" + _portalBranchName + "]");
-
-		Set<String> testBatchNames = getSetFromString(testBatchNamesString);
+		Set<String> testBatchNames = super.getBatchNames();
 
 		testBatchNames.addAll(_getOptionalBatchNames());
 
 		return testBatchNames;
-	}
-
-	@Override
-	public Set<String> getDependentBatchNames() {
-		String testBatchNames = JenkinsResultsParserUtil.getProperty(
-			jobProperties, "test.batch.names.smoke[" + _portalBranchName + "]");
-
-		return getSetFromString(testBatchNames);
-	}
-
-	@Override
-	public Set<String> getDistTypes() {
-		return Collections.emptySet();
-	}
-
-	@Override
-	public PortalGitWorkingDirectory getPortalGitWorkingDirectory() {
-		return _portalGitWorkingDirectory;
 	}
 
 	public void setPortalReleaseRef(String portalReleaseRef) {
@@ -96,9 +62,6 @@ public class PortalReleaseJob
 		return getSetFromString(testBatchNamesString);
 	}
 
-	private final GitWorkingDirectory _jenkinsGitWorkingDirectory;
-	private final String _portalBranchName;
-	private final PortalGitWorkingDirectory _portalGitWorkingDirectory;
 	private String _portalReleaseRef;
 
 }
